// Edu-Alt-Tech-website-main/types.ts
export type CourseCategory = 'education' | 'alternative';

export interface Course {
  id: string;
  title: string;
  description: string;
  category: CourseCategory;
  price?: number; // in INR, 0 = free
  thumbnailUrl?: string; // Optional thumbnail
  createdAt: any;
  createdBy: string; // admin or system
}